% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadVocabulary.R
\name{createCDMTables}
\alias{createCDMTables}
\title{Generate and execute the DDL on a database}
\usage{
createCDMTables(connectionDetails, cdmVersion, cdmDatabaseSchema, ...)
}
\arguments{
\item{connectionDetails}{An object of class connectionDetails as created by the DatabaseConnector::createConnectionDetails function.}

\item{cdmVersion}{The version of the CDM you are creating, e.g. 5.3, 5.4}

\item{cdmDatabaseSchema}{The schema of the CDM instance where the DDL will be run. For example, this would be "ohdsi.dbo" when testing on sql server.}

\item{...}{Other arguments passed on to DatabaseConnector::executeSql. (This allows the user to set the path to errorReportFile.)}
}
\description{
createCDMTables Creates CDM tables
}
\details{
This function will generate the DDL for a specific dbms and CDM version and
then execute the DDL on a database.
}
