% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadVocabulary.R
\name{dropConstraints}
\alias{dropConstraints}
\title{Drops OMOP CDM constraints}
\usage{
dropConstraints(connectionDetails, cdmVersion, cdmDatabaseSchema, ...)
}
\description{
Drops OMOP CDM constraints (foreign keys, primary keys, and indexes)
}
