% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatDdl.R
\name{formatCreateTables}
\alias{formatCreateTables}
\alias{formatCreatePrimaryKeys}
\alias{formatCreateForeignKeys}
\alias{formatCreateIndexes}
\alias{formatDropTables}
\alias{formatDropPrimaryKeys}
\alias{formatDropForeignKeys}
\alias{formatDropIndexes}
\alias{formatTruncateVocab}
\title{Create scripts to drop the OMOP CDM objects.}
\usage{
formatCreateTables(cdmVersion)

formatCreatePrimaryKeys(cdmVersion)

formatCreateForeignKeys(cdmVersion)

formatCreateIndexes(cdmVersion)

formatDropTables(cdmVersion)

formatDropPrimaryKeys(cdmVersion)

formatDropForeignKeys(cdmVersion)

formatDropIndexes(cdmVersion)

formatTruncateVocab(cdmVersion)
}
\arguments{
\item{cdmVersion}{The version of the CDM you are creating, e.g. 5.3, 5.4}
}
\value{
A character string containing the OHDSQL DDL
}
\description{
The formatDropTables, formatDropForeignKeys, and formatDropPrimaryKeys functions each return a character string
containing their respective DDL SQL code in OHDSQL dialect for a specific CDM version.
The SQL they generate needs to be rendered and translated before it can be executed.
}
\details{
The DDL SQL code is created from a two csv files that detail the OMOP CDM Specifications.
These files also form the basis of the CDM documentation and the Data Quality
Dashboard.
}
\section{Functions}{
\itemize{
\item \code{formatCreateTables}: formatCreateTables Returns a string containing the OHDSQL for creation of tables in the OMOP CDM.

\item \code{formatCreatePrimaryKeys}: v Returns a string containing the OHDSQL for creation of primary keys in the OMOP CDM.

\item \code{formatCreateForeignKeys}: formatCreateForeignKeys Returns a string containing the OHDSQL for creation of foreign keys in the OMOP CDM.

\item \code{formatCreateIndexes}: formatCreateIndexes Returns a string containing the OHDSQL for creation of indexes in the OMOP CDM.

\item \code{formatDropTables}: formatDropTables Returns a string containing the OHDSQL for deletion of primary keys in the OMOP CDM.

\item \code{formatDropPrimaryKeys}: formatDropPrimaryKeys Returns a string containing the OHDSQL for deletion of primary keys in the OMOP CDM.

\item \code{formatDropForeignKeys}: formatDropForeignKeys Returns a string containing the OHDSQL for deletion of foreign keys in the OMOP CDM.

\item \code{formatDropIndexes}: formatDropIndexes Returns a string containing the OHDSQL for deletion of indexes in the OMOP CDM.

\item \code{formatTruncateVocab}: formatTruncateVocab Returns a string containing the OHDSQL for truncation of vocabulary tables in the OMOP CDM.
}}

