% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadVocabulary.R
\name{loadVocabTables}
\alias{loadVocabTables}
\title{Loads the OMOP CDM vocabulary tables from csv files}
\usage{
loadVocabTables(
  connectionDetails,
  cdmVersion,
  cdmSchema,
  vocabFileLoc,
  bulkLoad = FALSE,
  chunkSize = 10000
)
}
\arguments{
\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using the
function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cdmVersion}{The version of the CDM you are creating, e.g. 5.3, 5.4}

\item{cdmSchema}{The name of the database schema that will contain the Vocabulary (and CDM)
tables.  Requires read and write permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'cdm_instance.dbo'.}

\item{vocabFileLoc}{The location of the vocabulary csv files.}

\item{chuckSize}{Number of records in each chunck of data that is processed.  Default is 10000}
}
\description{
Loads the OMOP CDM vocabulary tables from csv files
}
