% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testReadr.R
\name{processFile}
\alias{processFile}
\title{Tests processing vocabulary csv files}
\usage{
processFile(cdmVersion, fileLoc, chunkSize = 10000)
}
\description{
processFile Function to process vocabulary csv filessss.
}
