% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadVocabulary.R
\name{reapplyConstraints}
\alias{reapplyConstraints}
\title{Reapplies constraints on OMOP CDM vocabulary tables.}
\usage{
reapplyConstraints(connectionDetails, cdmVersion, cdmDatabaseSchema, ...)
}
\description{
Reapplies constraints on OMOP CDM vocabulary tables.
}
\section{Functions}{
\itemize{
\item \code{reapplyConstraints}: constraints on OMOP CDM vocabulary tables.
}}

