% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadVocabulary.R
\name{truncateVocabTables}
\alias{truncateVocabTables}
\title{Truncates OMOP CDM vocabulary tables}
\usage{
truncateVocabTables(connectionDetails, cdmVersion, cdmDatabaseSchema, ...)
}
\description{
Truncates OMOP CDM vocabulary tables
}
\section{Functions}{
\itemize{
\item \code{truncateVocabTables}: OMOP CDM vocabulary tables
}}

